Now able to send an RPC from the browser to the server. In the RPCPlugin I changed
  if (aParametersType != null) {
   boolean methodMatch = true;
   // TODO: added by Alex 2010-10-07: "&& j < lParametersType.length"
   for (int j = 0; j < aParametersType.length && j < lParametersType.length; j++) {
    if (!TypeConverter.matchProtocolTypeToJavaType(aParametersType[j], lParametersType[j].getName())) {
     methodMatch = false;
     break;
    }
   }

before that change it crashed with a ArrayIndexOutOfBoundException.
I intentionally didn't assign the right for rrpcTest3 and got a "method not available". Here please return a "access denied" and only if the method really does not exist a "method does not exist". I further updated the SampleRPCLibrary class

public String rrpcTest1() {
  logInfo("rrpcTest1");
  return "This is the result of 'rrpcTest1'.";
 }

to use the standard log4j logging and return a string a a result - and it was returned to the browser.